Build Form to edit Main Information
In the `ModifyRecipeView` added a picker to allow the user to modify a
different part of the recipe: Main Information, Ingredients, or
Directions.

Added a new view `ModifyMainInformationView` with a Form that allows the
user to modify the main information of a recipe. Added also colours to
the Form, to match the application's colour scheme.

This new Form is displayed when "Main Info" is selected on
`ModifyRecipeView`.