Build views to add ingredients to a recipe
Added a new view `ModifyIngredientView` to modify one ingredient.

Added a new view `ModifyIngredientsView` to modify multiple ingredients.
This view displays all the ingredients of a recipe.

If there are no ingredients, then `ModifyIngredientsView` should display
a `NavigationLink` displaying “Add a new ingredient” that links to a
`ModifyIngredientView`.

If there are ingredients to display, `ModifyIngredientsView` should
display a `List` structure that contains a `ForEach` that displays a
text view for each ingredient. At the bottom of the screen, there should
be a `NavigationLink` with the text “Add another ingredient” that links
to the `ModifyIngredientView`.

Added a `Save` button which should add the new ingredient to the list of
ingredients, and dismiss the form.

Added a new `Ingredient` initialiser to create an empty ingredient.

Added colour to the new views to match the application's color theme.