Add form to edit a Direction
- Added `ModifyDirectionView.swift` to edit a Direction.

- Instead of creating a new `ModifyDirectionsView` to edit multiple
directions, which would have been very similar to
`ModifyIngredientsView`, renamed `ModifyIngredientsView` to
`ModifyComponentsView` that will display either Ingrediensts of
Directions.

- Created a `RecipeComponent` protocol that `Ingredient` and `Component`
comform to.

- Created a `ModifyComponentView` protocol that `ModifyDirectionView`
and `ModifyIngredientView` comform to.

- Added two type parameter to `ModifyComponentsView`: A `Component` that
comforms to `RecipeComponent`, and a `DestinationView` that comforms to
`ModifyComponentView`. This will allow `ModifyComponentsView` to use a
generic `DestinationView` to display the appropriate view for the
component.

- All these changes required a bit of refactoring and fixing compilation
errors.

- Display the correct component name, either Ingredient or Direction, in
`ModifyComponentsView`.